restructured gremlin-python package
