+++ gremlin-variant/src/main/jython/.gitignore
