+++ gremlin-variant/src/main/jython/gremlin_python/__init__.py
@@ -16,20 +16,21 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 '''
-from gremlin_python import PythonGraphTraversal
-from gremlin_python import PythonGraphTraversalSource
-from gremlin_python import __
-from gremlin_python import statics
-from gremlin_python import T
-from gremlin_python import Order
-from gremlin_python import Cardinality
-from gremlin_python import Column
-from gremlin_python import Direction
-from gremlin_python import Operator
-from gremlin_python import P
-from gremlin_python import Pop
-from gremlin_python import Scope
-from gremlin_python import Barrier
-from groovy_translator import GroovyTranslator
+from .gremlin_python import PythonGraphTraversal
+from .gremlin_python import PythonGraphTraversalSource
+from .gremlin_python import __
+from .gremlin_python import statics
+from .gremlin_python import T
+from .gremlin_python import Order
+from .gremlin_python import Cardinality
+from .gremlin_python import Column
+from .gremlin_python import Direction
+from .gremlin_python import Operator
+from .gremlin_python import P
+from .gremlin_python import Pop
+from .gremlin_python import Scope
+from .gremlin_python import Barrier
+from .groovy_translator import GroovyTranslator
+
 
 __author__ = 'Marko A. Rodriguez (http://markorodriguez.com)'
