+++ gremlin-variant/src/main/jython/gremlin_python/gremlin_driver/__init__.py
@@ -16,6 +16,8 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 '''
-from gremlin_rest_driver import RESTRemoteConnection
+from .gremlin_driver import RemoteConnection
+from .gremlin_driver import Traverser
+
 
 __author__ = 'Marko A. Rodriguez (http://markorodriguez.com)'
