+++ gremlin-variant/src/main/jython/gremlin_python/gremlin_python.py
@@ -18,6 +18,8 @@ under the License.
 '''
 from collections import OrderedDict
 from aenum import Enum
+
+
 statics = OrderedDict()
