+++ gremlin-variant/src/main/jython/gremlin_python/groovy_translator.py
@@ -20,10 +20,10 @@ under the License.
 import sys
 from aenum import Enum
 
-from gremlin_python import P
-from gremlin_python import Raw
-from gremlin_python import RawExpression
-from translator import Translator
+from .gremlin_python import P
+from .gremlin_python import Raw
+from .gremlin_python import RawExpression
+from .translator import Translator
 
 if sys.version_info.major > 2:
     long = int
