+++ gremlin-variant/src/main/jython/gremlin_python/translator.py
@@ -18,6 +18,7 @@ under the License.
 '''
 from abc import abstractmethod
 
+
 __author__ = 'Marko A. Rodriguez (http://markorodriguez.com)'
